commited

+++ bingo/src/create_numbers.rs
@@ -1,9 +1,9 @@
 use rand::seq::SliceRandom;
 use rand::thread_rng;
-pub fn main() -> Vec<u32> {
+pub fn main() -> Vec<u8> {
   let mut numbers = Vec::new();
   for i in 1..=75 {
-    numbers.push(i as u32);
+    numbers.push(i as u8);
   }
   let mut rng = thread_rng();
   numbers.shuffle(&mut rng);
